+++ hooks/src/lib/useCVCValidation.ts
@@ -1,5 +1,5 @@
 import { useState } from 'react';
-import { Validation } from './validate';
+import { Validation } from './cardDateValidate';
 
 type CardCVCName = 'CVC';
